backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Kevin He <haokunkevinhe@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
selenium = "^4.25.0"
beautifulsoup4 = "^4.12.3"
pandas = "^2.2.3"
sqlmodel = "^0.0.22"
requests = "^2.32.3"
bs4 = "^0.0.2"
sec-edgar-api = "^1.1.0"
ipython = "^8.29.0"
tqdm = "^4.67.0"
sqlalchemy = ">=2.0.0"
pydantic-settings = "^2.6.1"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"
email-validator = "^2.2.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.10.0"
psycopg2-binary = "^2.9.10"
psycopg2 = "^2.9.10"
python-dotenv = "^1.0.1"
jwt = "^1.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"